FIX: 함수 작성순서 및 카멜케이스 수정
